Refactoring services to extract from profiles/common
