+++ controllers/profiles/common/properties/application_test.go
@@ -17,23 +17,24 @@
  * under the License.
  */
 
-package common
+package properties
 
 import (
 	"context"
 	"fmt"
 	"testing"
 
+	"github.com/apache/incubator-kie-kogito-serverless-operator/api/metadata"
 	operatorapi "github.com/apache/incubator-kie-kogito-serverless-operator/api/v1alpha08"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 
 	"github.com/apache/incubator-kie-kogito-serverless-operator/controllers/discovery"
+	"github.com/apache/incubator-kie-kogito-serverless-operator/controllers/platform/services"
 
 	"github.com/magiconair/properties"
 
 	"github.com/stretchr/testify/assert"
 
-	"github.com/apache/incubator-kie-kogito-serverless-operator/api/metadata"
 	"github.com/apache/incubator-kie-kogito-serverless-operator/test"
 )
 
@@ -46,6 +47,13 @@ const (
 	myService2Address = "http://10.110.90.2:80"
 	myService3        = "my-service3"
 	myService3Address = "http://10.110.90.3:80"
+
+	jobServiceURLProperty                   = "mp.messaging.outgoing.kogito-job-service-job-request-events.url"
+	jobServiceKafkaSinkInjectionHealthCheck = `quarkus.smallrye-health.check."org.kie.kogito.jobs.service.messaging.http.health.knative.KSinkInjectionHealthCheck".enabled`
+	jobServiceStatusChangeEventsProperty    = "kogito.jobs-service.http.job-status-change-events"
+	jobServiceStatusChangeEventsURL         = "mp.messaging.outgoing.kogito-job-service-job-status-events-http.url"
+	jobServiceURLProtocol                   = "http"
+	jobServiceDataSourceReactiveURLProperty = "quarkus.datasource.reactive.url"
 )
 
 type mockCatalogService struct {
@@ -87,137 +95,6 @@ func Test_appPropertyHandler_WithUserPropertiesWithNoUserOverrides(t *testing.T)
 	assert.Equal(t, "false", generatedProps.GetString("org.kie.kogito.addons.knative.eventing.health-enabled", ""))
 	assert.Equal(t, "false", generatedProps.GetString("quarkus.devservices.enabled", ""))
 	assert.Equal(t, "false", generatedProps.GetString("quarkus.kogito.devservices.enabled", ""))
-}
-
-func Test_appPropertyHandler_WithServicesWithUserOverrides(t *testing.T) {
-	//try to override kogito.service.url and quarkus.http.port
-	userProperties := "property1=value1\nproperty2=value2\nquarkus.http.port=9090\nkogito.service.url=http://myUrl.override.com\nquarkus.http.port=9090"
-	ns := "default"
-	workflow := test.GetBaseSonataFlow(ns)
-	enabled := true
-	platform := test.GetBasePlatform()
-	platform.Namespace = ns
-	platform.Spec = operatorapi.SonataFlowPlatformSpec{
-		Services: operatorapi.ServicesPlatformSpec{
-			DataIndex: &operatorapi.ServiceSpec{
-				Enabled: &enabled,
-			},
-			JobService: &operatorapi.ServiceSpec{
-				Enabled: &enabled,
-			},
-		},
-	}
-
-	di := NewDataIndexService(platform)
-
-	props := NewAppPropertyHandler(workflow, platform).WithUserProperties(userProperties).Build()
-	generatedProps, propsErr := properties.LoadString(props)
-	assert.NoError(t, propsErr)
-	assert.Equal(t, 8, len(generatedProps.Keys()))
-	assert.Equal(t, "value1", generatedProps.GetString("property1", ""))
-	assert.Equal(t, "value2", generatedProps.GetString("property2", ""))
-	//kogito.service.url takes the user provided value since it's a default mutable property.
-	assert.Equal(t, "http://myUrl.override.com", generatedProps.GetString("kogito.service.url", ""))
-	//quarkus.http.port remains with the default value since it's immutable.
-	assert.Equal(t, "8080", generatedProps.GetString("quarkus.http.port", ""))
-	assert.Equal(t, "0.0.0.0", generatedProps.GetString("quarkus.http.host", ""))
-	assert.Equal(t, "false", generatedProps.GetString("org.kie.kogito.addons.knative.eventing.health-enabled", ""))
-	assert.Equal(t, "false", generatedProps.GetString("quarkus.devservices.enabled", ""))
-	assert.Equal(t, "false", generatedProps.GetString("quarkus.kogito.devservices.enabled", ""))
-	assert.Equal(t, "", generatedProps.GetString(dataIndexServiceUrlProperty, ""))
-	assert.Equal(t, "", generatedProps.GetString(jobServiceURLProperty, ""))
-
-	// prod profile enables config of outgoing events url
-	workflow.SetAnnotations(map[string]string{metadata.Profile: string(metadata.ProdProfile)})
-	props = NewAppPropertyHandler(workflow, platform).WithUserProperties(userProperties).Build()
-	generatedProps, propsErr = properties.LoadString(props)
-	assert.NoError(t, propsErr)
-	assert.Equal(t, 13, len(generatedProps.Keys()))
-	assert.Equal(t, "http://"+platform.Name+"-"+DataIndexServiceName+"."+platform.Namespace+"/processes", generatedProps.GetString(dataIndexServiceUrlProperty, ""))
-	assert.Equal(t, "http://"+platform.Name+"-"+JobServiceName+"."+platform.Namespace+"/v2/jobs/events", generatedProps.GetString(jobServiceURLProperty, ""))
-	assert.Equal(t, "false", generatedProps.GetString(jobServiceKafkaSinkInjectionHealthCheck, ""))
-	assert.Equal(t, "", generatedProps.GetString(jobServiceDataSourceReactiveURLProperty, ""))
-	assert.Equal(t, "true", generatedProps.GetString(jobServiceStatusChangeEventsProperty, ""))
-	assert.Equal(t, generatedProps.GetString(jobServiceStatusChangeEventsURL, ""), fmt.Sprintf("%s://%s.%s/jobs", dataIndexServiceUrlProtocol, di.GetServiceName(), platform.Namespace))
-
-	// disabling data index bypasses config of outgoing events url
-	platform.Spec.Services.DataIndex.Enabled = nil
-	props = NewAppPropertyHandler(workflow, platform).WithUserProperties(userProperties).Build()
-	generatedProps, propsErr = properties.LoadString(props)
-	assert.NoError(t, propsErr)
-	assert.Equal(t, 10, len(generatedProps.Keys()))
-	assert.Equal(t, "", generatedProps.GetString(dataIndexServiceUrlProperty, ""))
-	assert.Equal(t, "http://"+platform.Name+"-"+JobServiceName+"."+platform.Namespace+"/v2/jobs/events", generatedProps.GetString(jobServiceURLProperty, ""))
-	assert.Equal(t, "false", generatedProps.GetString(jobServiceKafkaSinkInjectionHealthCheck, ""))
-	assert.Equal(t, "", generatedProps.GetString(jobServiceStatusChangeEventsProperty, ""))
-	assert.Equal(t, "", generatedProps.GetString(jobServiceStatusChangeEventsURL, ""))
-
-	// disabling job service bypasses config of outgoing events url
-	platform.Spec.Services.JobService.Enabled = nil
-	props = NewAppPropertyHandler(workflow, platform).WithUserProperties(userProperties).Build()
-	generatedProps, propsErr = properties.LoadString(props)
-	assert.NoError(t, propsErr)
-	assert.Equal(t, 8, len(generatedProps.Keys()))
-	assert.Equal(t, "", generatedProps.GetString(dataIndexServiceUrlProperty, ""))
-	assert.Equal(t, "", generatedProps.GetString(jobServiceURLProperty, ""))
-	assert.Equal(t, "", generatedProps.GetString(jobServiceKafkaSinkInjectionHealthCheck, ""))
-	assert.Equal(t, "", generatedProps.GetString(jobServiceDataSourceReactiveURLProperty, ""))
-	assert.Equal(t, "", generatedProps.GetString(jobServiceStatusChangeEventsProperty, ""))
-	assert.Equal(t, "", generatedProps.GetString(jobServiceStatusChangeEventsURL, ""))
-
-	// check that service app properties are being properly set
-	props = NewServiceAppPropertyHandler(platform).WithUserProperties(userProperties).Build()
-	generatedProps, propsErr = properties.LoadString(props)
-	assert.NoError(t, propsErr)
-	assert.Equal(t, 9, len(generatedProps.Keys()))
-	assert.Equal(t, "false", generatedProps.GetString(kafkaSmallRyeHealthProperty, ""))
-	assert.Equal(t, "value1", generatedProps.GetString("property1", ""))
-	assert.Equal(t, "value2", generatedProps.GetString("property2", ""))
-	//quarkus.http.port remains with the default value since it's immutable.
-	assert.Equal(t, "8080", generatedProps.GetString("quarkus.http.port", ""))
-
-	// check that the reactive URL is generated from the postgreSQL JDBC URL when not provided
-	platform.Spec.Services.JobService = &operatorapi.ServiceSpec{
-		Enabled: &enabled,
-		Persistence: &operatorapi.PersistenceOptions{
-			PostgreSql: &operatorapi.PersistencePostgreSql{
-				ServiceRef: operatorapi.PostgreSqlServiceOptions{
-					Name: "jobs-service",
-				},
-			},
-		},
-	}
-	props = NewAppPropertyHandler(workflow, platform).WithUserProperties(userProperties).Build()
-	generatedProps, propsErr = properties.LoadString(props)
-	assert.NoError(t, propsErr)
-	assert.Equal(t, 11, len(generatedProps.Keys()))
-	assert.Equal(t, "", generatedProps.GetString(dataIndexServiceUrlProperty, ""))
-	assert.Equal(t, "http://"+platform.Name+"-"+JobServiceName+"."+platform.Namespace+"/v2/jobs/events", generatedProps.GetString(jobServiceURLProperty, ""))
-	assert.Equal(t, "false", generatedProps.GetString(jobServiceKafkaSinkInjectionHealthCheck, ""))
-	assert.Equal(t, "postgresql://jobs-service.default:5432/sonataflow?search_path=sonataflow-platform-jobs-service", generatedProps.GetString(jobServiceDataSourceReactiveURLProperty, ""))
-	assert.Equal(t, "", generatedProps.GetString(jobServiceStatusChangeEventsProperty, ""))
-	assert.Equal(t, "", generatedProps.GetString(jobServiceStatusChangeEventsURL, ""))
-
-	// check that the reactive URL is generated from the postgreSQL JDBC URL when provided
-	platform.Spec.Services.JobService = &operatorapi.ServiceSpec{
-		Enabled: &enabled,
-		Persistence: &operatorapi.PersistenceOptions{
-			PostgreSql: &operatorapi.PersistencePostgreSql{
-				JdbcUrl: "jdbc:postgresql://timeouts-showcase-database:5432/postgres?currentSchema=jobs-service",
-			},
-		},
-	}
-	props = NewAppPropertyHandler(workflow, platform).WithUserProperties(userProperties).Build()
-	generatedProps, propsErr = properties.LoadString(props)
-	assert.NoError(t, propsErr)
-	assert.Equal(t, 11, len(generatedProps.Keys()))
-	assert.Equal(t, "", generatedProps.GetString(dataIndexServiceUrlProperty, ""))
-	assert.Equal(t, "http://"+platform.Name+"-"+JobServiceName+"."+platform.Namespace+"/v2/jobs/events", generatedProps.GetString(jobServiceURLProperty, ""))
-	assert.Equal(t, "false", generatedProps.GetString(jobServiceKafkaSinkInjectionHealthCheck, ""))
-	assert.Equal(t, "postgresql://timeouts-showcase-database:5432/postgres?search_path=jobs-service", generatedProps.GetString(jobServiceDataSourceReactiveURLProperty, ""))
-	assert.Equal(t, "", generatedProps.GetString(jobServiceStatusChangeEventsProperty, ""))
-	assert.Equal(t, "", generatedProps.GetString(jobServiceStatusChangeEventsURL, ""))
-
 }
 
 func Test_appPropertyHandler_WithUserPropertiesWithServiceDiscovery(t *testing.T) {
@@ -307,4 +184,135 @@ func Test_generateMicroprofileServiceCatalogProperty(t *testing.T) {
 func doTestGenerateMicroprofileServiceCatalogProperty(t *testing.T, serviceUri string, expectedProperty string) {
 	mpProperty := generateMicroprofileServiceCatalogProperty(serviceUri)
 	assert.Equal(t, mpProperty, expectedProperty, "expected microprofile service catalog property for serviceUri: %s, is %s, but the returned value was: %s", serviceUri, expectedProperty, mpProperty)
+}
+
+func Test_appPropertyHandler_WithServicesWithUserOverrides(t *testing.T) {
+	//try to override kogito.service.url and quarkus.http.port
+	userProperties := "property1=value1\nproperty2=value2\nquarkus.http.port=9090\nkogito.service.url=http://myUrl.override.com\nquarkus.http.port=9090"
+	ns := "default"
+	workflow := test.GetBaseSonataFlow(ns)
+	enabled := true
+	platform := test.GetBasePlatform()
+	platform.Namespace = ns
+	platform.Spec = operatorapi.SonataFlowPlatformSpec{
+		Services: operatorapi.ServicesPlatformSpec{
+			DataIndex: &operatorapi.ServiceSpec{
+				Enabled: &enabled,
+			},
+			JobService: &operatorapi.ServiceSpec{
+				Enabled: &enabled,
+			},
+		},
+	}
+
+	di := services.NewDataIndexService(platform)
+
+	props := NewAppPropertyHandler(workflow, platform).WithUserProperties(userProperties).Build()
+	generatedProps, propsErr := properties.LoadString(props)
+	assert.NoError(t, propsErr)
+	assert.Equal(t, 8, len(generatedProps.Keys()))
+	assert.Equal(t, "value1", generatedProps.GetString("property1", ""))
+	assert.Equal(t, "value2", generatedProps.GetString("property2", ""))
+	//kogito.service.url takes the user provided value since it's a default mutable property.
+	assert.Equal(t, "http://myUrl.override.com", generatedProps.GetString("kogito.service.url", ""))
+	//quarkus.http.port remains with the default value since it's immutable.
+	assert.Equal(t, "8080", generatedProps.GetString("quarkus.http.port", ""))
+	assert.Equal(t, "0.0.0.0", generatedProps.GetString("quarkus.http.host", ""))
+	assert.Equal(t, "false", generatedProps.GetString("org.kie.kogito.addons.knative.eventing.health-enabled", ""))
+	assert.Equal(t, "false", generatedProps.GetString("quarkus.devservices.enabled", ""))
+	assert.Equal(t, "false", generatedProps.GetString("quarkus.kogito.devservices.enabled", ""))
+	assert.Equal(t, "", generatedProps.GetString(services.DataIndexServiceUrlProperty, ""))
+	assert.Equal(t, "", generatedProps.GetString(jobServiceURLProperty, ""))
+
+	// prod profile enables config of outgoing events url
+	workflow.SetAnnotations(map[string]string{metadata.Profile: string(metadata.ProdProfile)})
+	props = NewAppPropertyHandler(workflow, platform).WithUserProperties(userProperties).Build()
+	generatedProps, propsErr = properties.LoadString(props)
+	assert.NoError(t, propsErr)
+	assert.Equal(t, 13, len(generatedProps.Keys()))
+	assert.Equal(t, "http://"+platform.Name+"-"+services.DataIndexServiceName+"."+platform.Namespace+"/processes", generatedProps.GetString(services.DataIndexServiceUrlProperty, ""))
+	assert.Equal(t, "http://"+platform.Name+"-"+services.JobServiceName+"."+platform.Namespace+"/v2/jobs/events", generatedProps.GetString(jobServiceURLProperty, ""))
+	assert.Equal(t, "false", generatedProps.GetString(jobServiceKafkaSinkInjectionHealthCheck, ""))
+	assert.Equal(t, "", generatedProps.GetString(jobServiceDataSourceReactiveURLProperty, ""))
+	assert.Equal(t, "true", generatedProps.GetString(jobServiceStatusChangeEventsProperty, ""))
+	assert.Equal(t, generatedProps.GetString(jobServiceStatusChangeEventsURL, ""), fmt.Sprintf("%s://%s.%s/jobs", services.DataIndexServiceUrlProtocol, di.GetServiceName(), platform.Namespace))
+
+	// disabling data index bypasses config of outgoing events url
+	platform.Spec.Services.DataIndex.Enabled = nil
+	props = NewAppPropertyHandler(workflow, platform).WithUserProperties(userProperties).Build()
+	generatedProps, propsErr = properties.LoadString(props)
+	assert.NoError(t, propsErr)
+	assert.Equal(t, 10, len(generatedProps.Keys()))
+	assert.Equal(t, "", generatedProps.GetString(services.DataIndexServiceUrlProperty, ""))
+	assert.Equal(t, "http://"+platform.Name+"-"+services.JobServiceName+"."+platform.Namespace+"/v2/jobs/events", generatedProps.GetString(jobServiceURLProperty, ""))
+	assert.Equal(t, "false", generatedProps.GetString(jobServiceKafkaSinkInjectionHealthCheck, ""))
+	assert.Equal(t, "", generatedProps.GetString(jobServiceStatusChangeEventsProperty, ""))
+	assert.Equal(t, "", generatedProps.GetString(jobServiceStatusChangeEventsURL, ""))
+
+	// disabling job service bypasses config of outgoing events url
+	platform.Spec.Services.JobService.Enabled = nil
+	props = NewAppPropertyHandler(workflow, platform).WithUserProperties(userProperties).Build()
+	generatedProps, propsErr = properties.LoadString(props)
+	assert.NoError(t, propsErr)
+	assert.Equal(t, 8, len(generatedProps.Keys()))
+	assert.Equal(t, "", generatedProps.GetString(services.DataIndexServiceUrlProperty, ""))
+	assert.Equal(t, "", generatedProps.GetString(jobServiceURLProperty, ""))
+	assert.Equal(t, "", generatedProps.GetString(jobServiceKafkaSinkInjectionHealthCheck, ""))
+	assert.Equal(t, "", generatedProps.GetString(jobServiceDataSourceReactiveURLProperty, ""))
+	assert.Equal(t, "", generatedProps.GetString(jobServiceStatusChangeEventsProperty, ""))
+	assert.Equal(t, "", generatedProps.GetString(jobServiceStatusChangeEventsURL, ""))
+
+	// check that service app properties are being properly set
+	props = NewServiceAppPropertyHandler(platform).WithUserProperties(userProperties).Build()
+	generatedProps, propsErr = properties.LoadString(props)
+	assert.NoError(t, propsErr)
+	assert.Equal(t, 9, len(generatedProps.Keys()))
+	assert.Equal(t, "false", generatedProps.GetString(KafkaSmallRyeHealthProperty, ""))
+	assert.Equal(t, "value1", generatedProps.GetString("property1", ""))
+	assert.Equal(t, "value2", generatedProps.GetString("property2", ""))
+	//quarkus.http.port remains with the default value since it's immutable.
+	assert.Equal(t, "8080", generatedProps.GetString("quarkus.http.port", ""))
+
+	// check that the reactive URL is generated from the postgreSQL JDBC URL when not provided
+	platform.Spec.Services.JobService = &operatorapi.ServiceSpec{
+		Enabled: &enabled,
+		Persistence: &operatorapi.PersistenceOptions{
+			PostgreSql: &operatorapi.PersistencePostgreSql{
+				ServiceRef: operatorapi.PostgreSqlServiceOptions{
+					Name: "jobs-service",
+				},
+			},
+		},
+	}
+	props = NewAppPropertyHandler(workflow, platform).WithUserProperties(userProperties).Build()
+	generatedProps, propsErr = properties.LoadString(props)
+	assert.NoError(t, propsErr)
+	assert.Equal(t, 11, len(generatedProps.Keys()))
+	assert.Equal(t, "", generatedProps.GetString(services.DataIndexServiceUrlProperty, ""))
+	assert.Equal(t, "http://"+platform.Name+"-"+services.JobServiceName+"."+platform.Namespace+"/v2/jobs/events", generatedProps.GetString(jobServiceURLProperty, ""))
+	assert.Equal(t, "false", generatedProps.GetString(jobServiceKafkaSinkInjectionHealthCheck, ""))
+	assert.Equal(t, "postgresql://jobs-service.default:5432/sonataflow?search_path=sonataflow-platform-jobs-service", generatedProps.GetString(jobServiceDataSourceReactiveURLProperty, ""))
+	assert.Equal(t, "", generatedProps.GetString(jobServiceStatusChangeEventsProperty, ""))
+	assert.Equal(t, "", generatedProps.GetString(jobServiceStatusChangeEventsURL, ""))
+
+	// check that the reactive URL is generated from the postgreSQL JDBC URL when provided
+	platform.Spec.Services.JobService = &operatorapi.ServiceSpec{
+		Enabled: &enabled,
+		Persistence: &operatorapi.PersistenceOptions{
+			PostgreSql: &operatorapi.PersistencePostgreSql{
+				JdbcUrl: "jdbc:postgresql://timeouts-showcase-database:5432/postgres?currentSchema=jobs-service",
+			},
+		},
+	}
+	props = NewAppPropertyHandler(workflow, platform).WithUserProperties(userProperties).Build()
+	generatedProps, propsErr = properties.LoadString(props)
+	assert.NoError(t, propsErr)
+	assert.Equal(t, 11, len(generatedProps.Keys()))
+	assert.Equal(t, "", generatedProps.GetString(services.DataIndexServiceUrlProperty, ""))
+	assert.Equal(t, "http://"+platform.Name+"-"+services.JobServiceName+"."+platform.Namespace+"/v2/jobs/events", generatedProps.GetString(jobServiceURLProperty, ""))
+	assert.Equal(t, "false", generatedProps.GetString(jobServiceKafkaSinkInjectionHealthCheck, ""))
+	assert.Equal(t, "postgresql://timeouts-showcase-database:5432/postgres?search_path=jobs-service", generatedProps.GetString(jobServiceDataSourceReactiveURLProperty, ""))
+	assert.Equal(t, "", generatedProps.GetString(jobServiceStatusChangeEventsProperty, ""))
+	assert.Equal(t, "", generatedProps.GetString(jobServiceStatusChangeEventsURL, ""))
+
 }
